Components structure refactored.
